+++ src/main/java/balancetalk/bookmark/domain/BookmarkRepository.java
@@ -1,0 +1,6 @@
+package balancetalk.bookmark.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BookmarkRepository extends JpaRepository<Bookmark, Long> {
+}
