--- src/main/java/balancetalk/bookmark/domain/repository/BookmarkRepository.java
@@ -1,7 +1,0 @@
-package balancetalk.bookmark.domain.repository;
-
-import balancetalk.bookmark.domain.Bookmark;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BookmarkRepository extends JpaRepository<Bookmark, Long>, BookmarkRepositoryCustom {
-}
