--- src/main/java/balancetalk/bookmark/domain/repository/BookmarkRepositoryCustom.java
@@ -1,8 +1,0 @@
-package balancetalk.bookmark.domain.repository;
-
-import balancetalk.bookmark.domain.BookmarkType;
-
-public interface BookmarkRepositoryCustom {
-
-    Long countBookmarksByResourceIdAndType(long resourceId, BookmarkType type);
-}
